calendar/cmd to calendar/service refactor
