+++ hw12_13_14_15_calendar/cmd/calendar/service/root.go
@@ -1,4 +1,4 @@
-package cmd
+package service
 
 import (
 	"encoding/json"
@@ -7,6 +7,8 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"github.com/xkdgo/otus_golang_homeworks/hw12_13_14_15_calendar/cmd/calendar/jobs/migrations"
+	"github.com/xkdgo/otus_golang_homeworks/hw12_13_14_15_calendar/internal/config"
 )
 
 // rootCmd represents the base command when called without any subcommands.
@@ -22,7 +24,7 @@ var (
 it can use several storages
 all configurations you can find in config file example`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			config, err := NewConfig(CfgFile)
+			config, err := config.NewConfig(CfgFile)
 			if err != nil {
 				return err
 			}
@@ -51,6 +53,13 @@ all configurations you can find in config file example`,
 			}
 		},
 	}
+
+	migrateCmd = &cobra.Command{
+		Use:   "migrate",
+		Short: "migrate cmd is used for database migration",
+		Long:  `migrate cmd is used for database migration: migrate <up | down>`,
+		RunE:  migrations.Migrate,
+	}
 )
 
 // Execute adds all child commands to the root command and sets flags appropriately.
@@ -61,6 +70,10 @@ func Execute() {
 	}
 }
 
+func GetRootCmd() *cobra.Command {
+	return rootCmd
+}
+
 func init() {
 	// Here you will define your flags and configuration settings.
 	// Cobra supports persistent flags, which, if defined here,
@@ -69,4 +82,7 @@ func init() {
 	viper.BindPFlag("config", rootCmd.Flags().Lookup("config"))
 
 	rootCmd.AddCommand(cmdVersion)
+	migrateCmd.Flags().StringVarP(&migrations.CfgFile, "config", "c", "/etc/calendar/config.toml", "Path to configuration file")
+	viper.BindPFlag("config", migrateCmd.Flags().Lookup("config"))
+	rootCmd.AddCommand(migrateCmd)
 }
