+++ hw12_13_14_15_calendar/cmd/calendar/version.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/xkdgo/otus_golang_homeworks/hw12_13_14_15_calendar/cmd/calendar/cmd"
+import "github.com/xkdgo/otus_golang_homeworks/hw12_13_14_15_calendar/cmd/calendar/service"
 
 var (
 	release   = "UNKNOWN"
@@ -9,7 +9,7 @@ var (
 )
 
 func transferVersionToCalendar() {
-	cmd.Release = release
-	cmd.BuildDate = buildDate
-	cmd.GitHash = gitHash
+	service.Release = release
+	service.BuildDate = buildDate
+	service.GitHash = gitHash
 }
