+++ hw12_13_14_15_calendar/internal/config/config.go
@@ -1,4 +1,4 @@
-package cmd
+package config
 
 import (
 	"fmt"
